switch: fix audio, remove audio option, small misc stuff
something is preventing me from running anything on the fourth core.
even the first build is broken now, and it was definitely working when i released it.
so i've moved the audio thread to the first core.
i also removed the audio option because it seems to have no impact on performance.
(does anyone want it as a feature? seems useless since you can just mute the system)
misc stuff includes a fix for the file brower backing out of the sd card.